First part of cleanup of position number and length usage - added typedefs to cinelerra/datatype.h for integer   position of audio sample, frame and abstract position - assuming that int is at least 32 bits - assuming that read/write length does fit in int - assuming that a memory buffer length does fit in int Removed some unused variables

+++ cinelerra/ffmpeg.C
@@ -365,7 +365,7 @@ int FFMPEG::convert_cmodel(AVPicture *picture_in, PixelFormat pix_fmt_in,
 	return 0;
 }
 
-int FFMPEG::decode(uint8_t *data, long data_size, VFrame *frame_out) { 
+int FFMPEG::decode(uint8_t *data, int data_size, VFrame *frame_out) { 
 
 	// NOTE: frame must already have data space allocated
